refactor(perl-lexer): split lexer unit tests into srp submodule (#0000)
